Organized Model file structure
