+++ composer.json
@@ -36,7 +36,8 @@
         },
         "classmap": [
             "database/seeds",
-            "database/factories"
+            "database/factories",
+            "app/Models"
         ]
     },
     "autoload-dev": {
